Simplify code with support methods.
