+++ module/VuFind/tests/integration-tests/src/VuFindTest/Mink/AutocompleteTest.php
@@ -29,10 +29,6 @@
 
 namespace VuFindTest\Mink;
 
-use Behat\Mink\Element\Element;
-use Behat\Mink\Element\NodeElement;
-use Behat\Mink\Session;
-
 /**
  * Mink test class for autocomplete functionality.
  *
@@ -45,46 +41,6 @@ use Behat\Mink\Session;
 class AutocompleteTest extends \VuFindTest\Integration\MinkTestCase
 {
     use \VuFindTest\Feature\AutocompleteTrait;
-
-    /**
-     * Load the Search/Home page as a foundation for searching.
-     *
-     * @param ?Session $session Mink session (will be automatically established if not provided).
-     *
-     * @return Element
-     */
-    protected function loadSearchHome(?Session $session = null): Element
-    {
-        $session ??= $this->getMinkSession();
-        $session->visit($this->getVuFindUrl() . '/Search/Home');
-        return $session->getPage();
-    }
-
-    /**
-     * For the provided search, assert the first autocomplete value and return the
-     * associated page element.
-     *
-     * @param string   $search   Search term(s)
-     * @param string   $expected First expected Autocomplete suggestion
-     * @param ?string  $type     Search type (null for default)
-     * @param ?Element $page     Existing page to use for searching (will load Search/Home if not provided)
-     *
-     * @return NodeElement
-     */
-    protected function assertAutocompleteValueAndReturnItem(
-        string $search,
-        string $expected,
-        ?string $type = null,
-        ?Element $page = null,
-    ): NodeElement {
-        $page ??= $this->loadSearchHome();
-        if ($type) {
-            $this->findCssAndSetValue($page, '#searchForm_type', $type);
-        }
-        $this->findCssAndSetValue($page, '#searchForm_lookfor', $search);
-        $acItem = $this->getAndAssertFirstAutocompleteValue($page, $expected);
-        return $acItem;
-    }
 
     /**
      * Test that default autocomplete behavior is correct.
@@ -155,7 +111,7 @@ class AutocompleteTest extends \VuFindTest\Integration\MinkTestCase
     {
         // Load the page first so we'll use the same context across all testing:
         $session = $this->getMinkSession();
-        $page = $this->loadSearchHome($session);
+        $page = $this->getSearchHomePage($session);
 
         // First do a search in All Fields
         $this->assertAutocompleteValueAndReturnItem('jsto', 'Al Gore', null, $page);
