+++ module/VuFind/tests/integration-tests/src/VuFindTest/Mink/LinkResolverTest.php
@@ -149,9 +149,7 @@ class LinkResolverTest extends \VuFindTest\Integration\MinkTestCase
         );
 
         // Search for a known record:
-        $session = $this->getMinkSession();
-        $session->visit($this->getVuFindUrl() . '/Search/Home');
-        $page = $session->getPage();
+        $page = $this->getSearchHomePage();
         $this->findCss($page, '#searchForm_lookfor')
             ->setValue('id:testsample1');
         $this->clickCss($page, '.btn.btn-primary');
@@ -177,12 +175,7 @@ class LinkResolverTest extends \VuFindTest\Integration\MinkTestCase
         );
 
         // Search for a known record:
-        $session = $this->getMinkSession();
-        $session->visit($this->getVuFindUrl() . '/Search/Home');
-        $page = $session->getPage();
-        $this->findCss($page, '#searchForm_lookfor')
-            ->setValue('id:testsample1');
-        $this->clickCss($page, '.btn.btn-primary');
+        $page = $this->performSearch('id:testsample1');
 
         // Verify the OpenURL
         $this->assertOpenUrl($page, false /* do not click link */);
